requirements.txt
transformers==4.29.2
accelerate==0.19.0
datasets
pysbd
wandb
h5py
nltk
spacy
ersatz
iso-639
scikit-learn==1.2.2
numpy==1.23.5
pydantic
torchinfo
conllu
genalog
pandarallel
cohere
replicate
onnx
onnxruntime
torchinfo